Use NativeEndian when parsing DM ioctl results
The data returned from ioctls will be in whatever the native endianness
of the machine is. We don't want to endian-convert, we just want to extract
multibyte values out of the [u8], so use NativeEndian instead of
LittleEndian and the right thing should happen.

This is in contrast with the on-disk format, which is little-endian.

Signed-off-by: Andy Grover <agrover@redhat.com>