Index PVs by Device instead of "pv0"-style string
"pv0"-style strings in the metadata are needed to identify the PVs that
an LV's strips is allocated from, but are not the best way to index PVs
within a VG. Instead, convert PV and stripe references to Device when
creating objects from a textmap, and convert back from Device to "pv0"
string when converting to a textmap.

To build LV textmaps we need to pass a map of Device->ordinal so it's no
longer possible to implement the From trait.

Also, we no longer move args into *::to_textmap() functions for the
various structs.

Signed-off-by: Andy Grover <agrover@redhat.com>